+++ common/core/src/main/java/is/codion/common/state/DefaultStateCombination.java
@@ -109,8 +109,8 @@ final class DefaultStateCombination implements State.Combination {
 	}
 
 	@Override
-	public boolean nullable() {
-		return observer.nullable();
+	public boolean isNullable() {
+		return observer.isNullable();
 	}
 
 	@Override