+++ swing/common-ui/src/main/java/is/codion/swing/common/ui/component/button/DefaultToggleButtonBuilder.java
@@ -38,7 +38,7 @@ class DefaultToggleButtonBuilder<C extends JToggleButton, B extends ToggleButton
 
 	@Override
 	public final B toggleControl(ToggleControl toggleControl) {
-		if (requireNonNull(toggleControl).value().nullable() && !supportsNull()) {
+		if (requireNonNull(toggleControl).value().isNullable() && !supportsNull()) {
 			throw new IllegalArgumentException("This toggle button builder does not support a nullable value");
 		}
 		this.toggleControl = toggleControl;