+++ swing/common-ui/src/main/java/is/codion/swing/common/ui/component/text/NumberField.java
@@ -623,7 +623,7 @@ public final class NumberField<T extends Number> extends HintTextField {
 		protected Short getComponentValue() {
 			Number number = component().getNumber();
 			if (number == null) {
-				return nullable() ? null : (short) 0;
+				return isNullable() ? null : (short) 0;
 			}
 
 			return number.shortValue();
@@ -645,7 +645,7 @@ public final class NumberField<T extends Number> extends HintTextField {
 		protected Double getComponentValue() {
 			Number number = component().getNumber();
 			if (number == null) {
-				return nullable() ? null : 0d;
+				return isNullable() ? null : 0d;
 			}
 
 			return number.doubleValue();
@@ -667,7 +667,7 @@ public final class NumberField<T extends Number> extends HintTextField {
 		protected Integer getComponentValue() {
 			Number number = component().getNumber();
 			if (number == null) {
-				return nullable() ? null : 0;
+				return isNullable() ? null : 0;
 			}
 
 			return number.intValue();
@@ -689,7 +689,7 @@ public final class NumberField<T extends Number> extends HintTextField {
 		protected Long getComponentValue() {
 			Number number = component().getNumber();
 			if (number == null) {
-				return nullable() ? null : 0L;
+				return isNullable() ? null : 0L;
 			}
 
 			return number.longValue();
@@ -711,7 +711,7 @@ public final class NumberField<T extends Number> extends HintTextField {
 		protected BigDecimal getComponentValue() {
 			BigDecimal number = component().getNumber();
 			if (number == null) {
-				return nullable() ? null : BigDecimal.ZERO;
+				return isNullable() ? null : BigDecimal.ZERO;
 			}
 
 			return number;