+++ changelog.md
@@ -7,6 +7,7 @@ Codion Change Log
 - Value.Builder, methods for adding listeners and consumers added.
 - ValueSet.Builder, methods for adding listeners and consumers added.
 - State.Builder added.
+- ValueObserver.nullable() renamed isNullable(), now consistent with other method names.
 ### is.codion.framework.domain
 - OrderBy.Builder, NullOrder now parameter.
 - OrderBy.OrderByColumn.ignoreCase() added.